refactor(user): Redis 관련 코드 재구조화 (#34)
- RedisToken 엔티티 이동 (infra/redis -> feature/auth/redis)
    - 비즈니스 핵심 규칙이 아니라 인증이라는 기능을 구현하기 위한
      기술적인 도구이므로 feature 패키지 아래에 위치
- RedisConfig 전역 설정으로 변경 (user/infra/redis/config -> api/config)
- 기존 user 모듈의 Redis 관련 테스트를 위한 전용 Config 클래스 생성 및
  주입 (TestRedistConfig)
