+++ api-server/build.gradle
@@ -18,4 +18,6 @@ dependencies {
 
     testImplementation 'com.tngtech.archunit:archunit-junit5:1.3.0'
     testImplementation 'jakarta.persistence:jakarta.persistence-api'
+
+    implementation 'org.springframework.boot:spring-boot-starter-data-redis'
 }
