+++ api-server/src/main/java/com/teno/openmarket/api/config/RedisConfig.java
@@ -1,4 +1,4 @@
-package com.teno.openmarket.user.infra.redis.config;
+package com.teno.openmarket.api.config;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
