+++ user/src/main/java/com/teno/openmarket/user/feature/auth/redis/RefreshToken.java
@@ -1,4 +1,4 @@
-package com.teno.openmarket.user.infra.redis.entity;
+package com.teno.openmarket.user.feature.auth.redis;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
